refactor: rename userCategories.ts and delete ? in front of selectedCategory
